[refactor]: PointUtil 디렉토리 변경 crossroad -> global
사유 : 여러 도메인에서 접근하는 Util 클래스
